Update backend to allow requesting candidate csv with rod data
