+++ api/graphql/schema.graphql
@@ -2396,6 +2396,7 @@ type Mutation {
   downloadPoolCandidatesCsv(
     ids: [UUID!]
     where: PoolCandidateSearchInput
+    withROD: Boolean
   ): Boolean! @guard
   downloadUsersCsv(ids: [UUID!], where: UserFilterInput): Boolean! @guard
