+++ api/lang/en/common.php
@@ -18,4 +18,5 @@ return [
     'present' => 'Present',
     'at' => 'at',
     'not_provided' => 'Not provided',
+    'not_sure' => 'Not sure',
 ];
